workspace/projects/ai-council-system/requirements.txt
# AI Council System - Requirements
# Install with: pip install -r requirements.txt

# Core dependencies
pyyaml>=6.0
python-dateutil>=2.8.0

# LLM Providers
anthropic>=0.18.0
openai>=1.12.0

# Event Sources
tweepy>=4.14.0  # Twitter API
newsapi-python>=0.2.7  # News API
feedparser>=6.0.10  # RSS feeds
requests>=2.31.0  # HTTP requests

# Text-to-Speech
elevenlabs>=0.2.26  # ElevenLabs TTS
pyttsx3>=2.90  # Offline TTS fallback
gTTS>=2.4.0  # Google TTS fallback

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
python-multipart>=0.0.6

# Streaming & Media
ffmpeg-python>=0.2.0
Pillow>=10.2.0
opencv-python>=4.9.0

# Data Processing
numpy>=1.24.0
pandas>=2.1.0

# Utilities
aiohttp>=3.9.0
aiofiles>=23.2.0
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development
black>=24.1.0
flake8>=7.0.0
mypy>=1.8.0
ipython>=8.20.0

# Blockchain (Phase 3)
web3>=6.15.0  # Ethereum/Web3
eth-account>=0.10.0  # Ethereum accounts
eth-utils>=2.3.0  # Ethereum utilities
solana>=0.30.0  # Solana Python client
anchorpy>=0.15.0  # Anchor framework Python bindings
solders>=0.18.0  # Solana SDK
base58>=2.1.1  # Base58 encoding
pynacl>=1.5.0  # Libsodium/NaCl crypto

# Avatar Generation (Phase 4.1)
replicate>=0.22.0  # Replicate API for Stable Diffusion
# openai>=1.12.0  # Already included above for DALL-E 3
# Pillow>=10.2.0  # Already included above for image processing
# numpy>=1.24.0  # Already included above
